requirements.txt
snowflake-connector-python==3.5.0
streamlit==1.32.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
